feat:finish game and view winners
