+++ package.json
@@ -2,7 +2,7 @@
   "name": "ws_task",
   "version": "1.0.0",
   "description": "Template for the WebSocket remote control task",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "start": "tsc && node dist/index.js",
     "start:dev": "nodemon ./index.ts",
