+++ src/webSocket_server/requestsHandlers/reg_handler.ts
@@ -1,5 +1,4 @@
 import { connections } from "../../dataBase/dataBaseConnections";
-import { dataBaseUsers } from "../../dataBase/dataBaseUsers";
 import { Player, Request } from "../../utils/types";
 import { WebSocket } from "ws";
 
@@ -12,11 +11,10 @@ export const regHandler = (ws: WebSocket, req: Request, connectionId: string) =>
     password: dataPlayer.password,
     idPlayer: connectionId,
     ws: ws,
+    wins: 0,
   } as Player;
 
   connections.set(connectionId, player);
-
-  dataBaseUsers.players.push(player); //????
 
   const resp = {
     type: "reg",
